WK 3 Part 1 Not Finished (Step 5)
We're stuck on Step 5 because I have no flipping clue what he's talking about in any of the instructions.

We're getting errors all over the place. He tells us to assign the variables in the blueprint but like how am I even supposed to do that?

I don't understand when and where things are meant to be being done because he's not explaining it well.

So we'll have to message him to tell him that we don't understand what step 5 means.