Week 3 P3 issue FIXED!
The problem was actually that the player was way too far down in the world, which caused him to automatically die; which is why we never saw the player or the environment.

He would just instantly die; so we fixed the issue by bringing the player up to y level 0, and dragging the tile map up to his level in order to keep him high enough so that he doesn't die from fall damage!