Still on WK4 PT 1 STEP 1
We're currently at this point

"As mentioned in the lecture, the constructor tells us how to build an object, but in Unreal, we also add an interface for us to work with.
Not for the collider use this to help set up:
https://dev.epicgames.com/documentation/en-us/unreal-engine/API/Runtime/Engine/Components/UBoxComponent
You don’t want to replace the RootComponent, but you can attach a child object like this:
YourCollider->SetupAttachment(RootComponent);"

This part is quite confusing. We'll figure it all out tomorrow when we're at the Dalton Building / Library.